// src/generated/models/Stf_strategypocsModel.ts
export const Stf_strategypocspm_startmonth = {
  1: 'January',
  2: 'February',
  3: 'March',
  4: 'April',
  5: 'May',
  6: 'June',
  7: 'July',
  8: 'August',
  9: 'September',
  10: 'October',
  11: 'November',
  12: 'December'
} as const;
export type Stf_strategypocspm_startmonth = keyof typeof Stf_strategypocspm_startmonth;
export const Stf_strategypocsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Stf_strategypocsstatecode = keyof typeof Stf_strategypocsstatecode;
export const Stf_strategypocsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Stf_strategypocsstatuscode = keyof typeof Stf_strategypocsstatuscode;
export const Stf_strategypocsstf_pocstatus = {
  1: 'Active',
  2: 'Succeeded',
  3: 'Failed',
  4: 'Retired'
} as const;
export type Stf_strategypocsstf_pocstatus = keyof typeof Stf_strategypocsstf_pocstatus;
export const Stf_strategypocsstf_serviceexecutionmode = {
  1: 'TMS Task',
  2: 'Project'
} as const;
export type Stf_strategypocsstf_serviceexecutionmode = keyof typeof Stf_strategypocsstf_serviceexecutionmode;
export const Stf_strategypocsstf_strategytype = {
  989230000: 'Department Strategy',
  989230001: 'Market Strategy',
  989230002: 'Specialty Strategy',
  989230003: 'Service Strategy'
} as const;
export type Stf_strategypocsstf_strategytype = keyof typeof Stf_strategypocsstf_strategytype;

export interface Stf_strategypocsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  "pm_Model@odata.bind"?: string;
  pm_startmonth?: Stf_strategypocspm_startmonth;
  statecode: Stf_strategypocsstatecode;
  statuscode?: Stf_strategypocsstatuscode;
  "stf_ConvertedTactic@odata.bind"?: string;
  stf_currentbaseline?: number;
  stf_experimentscope?: string;
  stf_from?: string;
  stf_killcondition?: string;
  stf_killconditiondate?: string;
  "stf_KPI@odata.bind"?: string;
  stf_kpitargetvalue?: number;
  stf_neededbudget?: number;
  "stf_POCCategory@odata.bind"?: string;
  stf_pocdescription?: string;
  stf_pocid?: string;
  stf_pocname?: string;
  stf_pocstatus?: Stf_strategypocsstf_pocstatus;
  "stf_Project@odata.bind"?: string;
  "stf_Region@odata.bind"?: string;
  stf_serviceexecutionmode?: Stf_strategypocsstf_serviceexecutionmode;
  "stf_Specialty@odata.bind"?: string;
  "stf_StrategyKPI@odata.bind"?: string;
  stf_strategypocid: string;
  stf_successcriteria?: string;
  stf_successduedate?: string;
  stf_target?: number;
  stf_to?: string;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Stf_strategypocs extends Stf_strategypocsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_modelname?: string;
  pm_startmonthname?: string;
  statecodename?: string;
  statuscodename?: string;
  stf_convertedtacticname?: string;
  stf_kpiname?: string;
  stf_poccategoryname?: string;
  stf_pocstatusname?: string;
  stf_projectname?: string;
  stf_regionname?: string;
  stf_serviceexecutionmodename?: string;
  stf_specialtyname?: string;
  stf_strategykpiname?: string;
  stf_strategytype?: Stf_strategypocsstf_strategytype;
  stf_strategytypename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pm_model?: object;
  _pm_model_value?: string;
  stf_convertedtactic?: object;
  _stf_convertedtactic_value?: string;
  stf_kpi?: object;
  _stf_kpi_value?: string;
  stf_poccategory?: object;
  _stf_poccategory_value?: string;
  stf_project?: object;
  _stf_project_value?: string;
  stf_region?: object;
  _stf_region_value?: string;
  stf_specialty?: object;
  _stf_specialty_value?: string;
  stf_strategykpi?: object;
  _stf_strategykpi_value?: string;
}